tests: move test_viva_provider to conftest.py
